Add checksum verification when loading an event log
When loading an event log, checksum verification is now enabled on the
loaded RDDs. Whenever a partition of a loaded RDD is recomputed, its
checksum is compared with the stored checksum. If there is a mismatch,
it is appended to EventLogReader.checksumMismatches.

This commit also fixes a race condition that occurred when calling
SparkContext.stop() while there were events waiting to be sent through
Akka to the master. Previously, EventReporter.stop() would stop the
EventLogWriter immediately, dropping those pending events. This commit
fixes the issue by stopping EventReporter using an Akka message.